refactor(cohort): split out cohortLogic into cohortLogic and cohortEditLogic
